+++ src/plugin/re.ts
@@ -5,7 +5,7 @@ import { RPCError } from "telegram/errors";
 const prefixes = getPrefixes();
 const mainPrefix = prefixes[0];
 class RePlugin extends Plugin {
-  ignoreEdited: boolean = false;
+  ignoreEdited: boolean = true;
   description: string = `复读\n回复一条消息即可复读\n<code>${mainPrefix}re [消息数] [复读次数]</code>`;
   cmdHandlers: Record<
     string,